yet more work on the tasks ystem
almost done i think

+++ gamemodes/efgmrewrite/gamemode/server/sv_wipe_manager.lua
@@ -1,6 +1,6 @@
 -- "formally" initializes a wipe, handling shit like resetting stats
 
-local wipeEnabled = false -- basically so nobody accidentally starts a wipe
+local wipeEnabled = false -- so nobody accidentally starts a wipe
 
 concommand.Add("efgm_unlock_wipe", function(ply, cmd, args)
 
@@ -20,12 +20,24 @@ end)
 
 concommand.Add("efgm_wipe_gameplay", function(ply, cmd, args) -- regular wipe, resets xp, level, stuff like that, keeps lifetime stats and unlocks
 
+    if !wipeEnabled then return end
+
+    print("gameplay wipe doesn't do anything yet")
+
+end)
+
+concommand.Add("efgm_wipe_tasks", function(ply, cmd, args)
+
+    if !wipeEnabled then return end
+
     print("gameplay wipe doesn't do anything yet")
 
 end)
 
 concommand.Add("efgm_wipe_full", function(ply, cmd, args) -- full reset of everything, should only be used in testing environments or MAYBE on a major enough update
 
+    if !wipeEnabled then return end
+    
     sql.Query("DROP TABLE EFGMPlayerData64; CREATE TABLE IF NOT EXISTS EFGMPlayerData64 ( SteamID INTEGER, Key TEXT, Value TEXT);")
 
     print("Did full wipe ig")
